fix square root calculation

+++ Project/src/common.c
@@ -17,6 +17,7 @@
  */
 
 #include "common.h"
+#include <math.h>
 
 void RCC_GPIOClockCmd(GPIO_TypeDef* GPIOx, FunctionalState state)
 {
@@ -221,5 +222,5 @@ void Timer_16bit_Calc(int freq, uint16_t *period, uint16_t *prescaler)
 
 float Distance2D(float dx, float dy)
 {
-	return sqrt(dx*dx + dy*dy);
+	return sqrtf(dx*dx + dy*dy);
 }
